+++ Changelog.md
@@ -1,3 +1,6 @@
+### ToyBox DarkHeresy - Ver 0.1.13 (built for 0.4.0.763)
+- (***ADDB***) Add Bag of Tricks => Other Multipliers => "Enemy Morale Multiplier"
+- (***ADDB***) Add Bag of Tricks => Other Multipliers => "Ally Morale Multiplier"
 - (***ADDB***) All 4 factions should now by shown in "Modify Faction Reputation" from the start.
 
 ### ToyBox DarkHeresy - Ver 0.1.12 (built for 0.4.0.763)